src/requirements.txt
nltk==3.9.1
numpy==1.26.4
gensim==4.3.3
pandas==2.2.3
scipy==1.15.3
scikit-learn==1.6.1
matplotlib==3.10.0
gitpython==3.1.44